+++ intellij-plugin-verifier/verifier-intellij/src/main/java/com/jetbrains/pluginverifier/usages/overrideOnly/AnActionUpdateMethodAllowedUsageFilter.kt
@@ -1,6 +1,7 @@
 package com.jetbrains.pluginverifier.usages.overrideOnly
 
 import com.jetbrains.pluginverifier.verifiers.filter.ApiUsageFilter
+import com.jetbrains.pluginverifier.verifiers.resolution.MethodDescriptor
 
 
 typealias BinaryClassName = String
