+++ intellij-plugin-verifier/verifier-intellij/src/main/java/com/jetbrains/pluginverifier/usages/overrideOnly/OverrideOnlyMethodAllowedUsageFilter.kt
@@ -4,7 +4,6 @@ import com.jetbrains.pluginverifier.verifiers.VerificationContext
 import com.jetbrains.pluginverifier.verifiers.filter.ApiUsageFilter
 import com.jetbrains.pluginverifier.verifiers.resolution.Method
 import org.objectweb.asm.tree.AbstractInsnNode
-import org.objectweb.asm.tree.MethodInsnNode
 
 class OverrideOnlyMethodAllowedUsageFilter : ApiUsageFilter {
 
@@ -21,11 +20,3 @@ class OverrideOnlyMethodAllowedUsageFilter : ApiUsageFilter {
     return isSuperCall || isDelegateCall
   }
 }
-
-data class MethodDescriptor(private val methodName: String, private val descriptor: String)
-
-fun Method.matches(method: Method): Boolean =
-  MethodDescriptor(name, descriptor) == MethodDescriptor(method.name, method.descriptor)
-
-fun MethodInsnNode.matches(method: Method): Boolean =
-  MethodDescriptor(name, desc) == MethodDescriptor(method.name, method.descriptor)
