+++ intellij-plugin-verifier/verifier-intellij/src/main/java/com/jetbrains/pluginverifier/usages/overrideOnly/SuperclassCallOnOverrideOnlyUsageFilter.kt
@@ -4,6 +4,7 @@ import com.jetbrains.pluginverifier.verifiers.VerificationContext
 import com.jetbrains.pluginverifier.verifiers.filter.ApiUsageFilter
 import com.jetbrains.pluginverifier.verifiers.resolution.Method
 import com.jetbrains.pluginverifier.verifiers.resolution.isCallOfSuperMethod
+import com.jetbrains.pluginverifier.verifiers.resolution.matches
 import org.objectweb.asm.tree.AbstractInsnNode
 
 class SuperclassCallOnOverrideOnlyUsageFilter  : ApiUsageFilter {
